Configure yt default player using a ~/.config/yt settings file
The user can set the default player in a ~/.config/yt settings file, by
reading the "player" entry, in the "[General]" section of the settings
file. The format of the settings file should be like this:

[General]
player = omxplayer

If the file is not there, or if it's corrupted in any way, yt defaults
to mplayer.